+++ public/app/features/dashboard/state/DashboardModel.test.ts
@@ -151,11 +151,6 @@ describe('DashboardModel', () => {
       expect(model.meta.canSave).toBe(false);
       expect(model.meta.canEdit).toBe(false);
     });
-
-    it('getSaveModelClone should remove meta', () => {
-      const clone = model.getSaveModelClone();
-      expect(clone.meta).toBe(undefined);
-    });
   });
 
   describe('when loading dashboard with old influxdb query schema', () => {
@@ -461,173 +456,6 @@ describe('DashboardModel', () => {
       it('should only remove row', () => {
         expect(dashboard.panels.length).toBe(4);
       });
-    });
-  });
-
-  describe('Given model with time', () => {
-    let model: DashboardModel;
-
-    beforeEach(() => {
-      model = new DashboardModel({
-        time: {
-          from: 'now-6h',
-          to: 'now',
-        },
-      });
-      expect(model.hasTimeChanged()).toBeFalsy();
-      model.time = {
-        from: 'now-3h',
-        to: 'now-1h',
-      };
-    });
-
-    it('hasTimeChanged should be true', () => {
-      expect(model.hasTimeChanged()).toBeTruthy();
-    });
-
-    it('getSaveModelClone should return original time when saveTimerange=false', () => {
-      const options = { saveTimerange: false };
-      const saveModel = model.getSaveModelClone(options);
-
-      expect(saveModel.time.from).toBe('now-6h');
-      expect(saveModel.time.to).toBe('now');
-    });
-
-    it('getSaveModelClone should return updated time when saveTimerange=true', () => {
-      const options = { saveTimerange: true };
-      const saveModel = model.getSaveModelClone(options);
-
-      expect(saveModel.time.from).toBe('now-3h');
-      expect(saveModel.time.to).toBe('now-1h');
-    });
-
-    it('hasTimeChanged should be false when reset original time', () => {
-      model.resetOriginalTime();
-      expect(model.hasTimeChanged()).toBeFalsy();
-    });
-
-    it('getSaveModelClone should return original time when saveTimerange=false', () => {
-      const options = { saveTimerange: false };
-      const saveModel = model.getSaveModelClone(options);
-
-      expect(saveModel.time.from).toBe('now-6h');
-      expect(saveModel.time.to).toBe('now');
-    });
-
-    it('getSaveModelClone should return updated time when saveTimerange=true', () => {
-      const options = { saveTimerange: true };
-      const saveModel = model.getSaveModelClone(options);
-
-      expect(saveModel.time.from).toBe('now-3h');
-      expect(saveModel.time.to).toBe('now-1h');
-    });
-
-    it('getSaveModelClone should remove repeated panels and scopedVars', () => {
-      const dashboardJSON = {
-        panels: [
-          { id: 1, type: 'row', repeat: 'dc', gridPos: { x: 0, y: 0, h: 1, w: 24 } },
-          { id: 2, repeat: 'app', repeatDirection: 'h', gridPos: { x: 0, y: 1, h: 2, w: 8 } },
-        ],
-        templating: {
-          list: [
-            {
-              name: 'dc',
-              type: 'custom',
-              current: {
-                text: 'dc1 + dc2',
-                value: ['dc1', 'dc2'],
-              },
-              options: [
-                { text: 'dc1', value: 'dc1', selected: true },
-                { text: 'dc2', value: 'dc2', selected: true },
-              ],
-            },
-            {
-              name: 'app',
-              type: 'custom',
-              current: {
-                text: 'se1 + se2',
-                value: ['se1', 'se2'],
-              },
-              options: [
-                { text: 'se1', value: 'se1', selected: true },
-                { text: 'se2', value: 'se2', selected: true },
-              ],
-            },
-          ],
-        },
-      };
-
-      const model = getDashboardModel(dashboardJSON);
-      model.processRepeats();
-      expect(model.panels.filter((x) => x.type === 'row')).toHaveLength(2);
-      expect(model.panels.filter((x) => x.type !== 'row')).toHaveLength(4);
-      expect(model.panels.find((x) => x.type !== 'row')?.scopedVars?.dc.value).toBe('dc1');
-      expect(model.panels.find((x) => x.type !== 'row')?.scopedVars?.app.value).toBe('se1');
-
-      const saveModel = model.getSaveModelClone();
-      expect(saveModel.panels.length).toBe(2);
-      expect(saveModel.panels[0].scopedVars).toBe(undefined);
-      expect(saveModel.panels[1].scopedVars).toBe(undefined);
-
-      model.collapseRows();
-      const savedModelWithCollapsedRows: any = model.getSaveModelClone();
-      expect(savedModelWithCollapsedRows.panels[0].panels.length).toBe(1);
-    });
-
-    it('getSaveModelClone should not remove repeated panels and scopedVars during snapshot', () => {
-      const dashboardJSON = {
-        panels: [
-          { id: 1, type: 'row', repeat: 'dc', gridPos: { x: 0, y: 0, h: 1, w: 24 } },
-          { id: 2, repeat: 'app', repeatDirection: 'h', gridPos: { x: 0, y: 1, h: 2, w: 8 } },
-        ],
-        templating: {
-          list: [
-            {
-              name: 'dc',
-              type: 'custom',
-              current: {
-                text: 'dc1 + dc2',
-                value: ['dc1', 'dc2'],
-              },
-              options: [
-                { text: 'dc1', value: 'dc1', selected: true },
-                { text: 'dc2', value: 'dc2', selected: true },
-              ],
-            },
-            {
-              name: 'app',
-              type: 'custom',
-              current: {
-                text: 'se1 + se2',
-                value: ['se1', 'se2'],
-              },
-              options: [
-                { text: 'se1', value: 'se1', selected: true },
-                { text: 'se2', value: 'se2', selected: true },
-              ],
-            },
-          ],
-        },
-      };
-
-      const model = getDashboardModel(dashboardJSON);
-      model.processRepeats();
-      expect(model.panels.filter((x) => x.type === 'row')).toHaveLength(2);
-      expect(model.panels.filter((x) => x.type !== 'row')).toHaveLength(4);
-      expect(model.panels.find((x) => x.type !== 'row')?.scopedVars?.dc.value).toBe('dc1');
-      expect(model.panels.find((x) => x.type !== 'row')?.scopedVars?.app.value).toBe('se1');
-
-      model.snapshot = { timestamp: new Date() };
-      const saveModel = model.getSaveModelClone();
-      expect(saveModel.panels.filter((x) => x.type === 'row')).toHaveLength(2);
-      expect(saveModel.panels.filter((x) => x.type !== 'row')).toHaveLength(4);
-      expect(saveModel.panels.find((x) => x.type !== 'row')?.scopedVars?.dc.value).toBe('dc1');
-      expect(saveModel.panels.find((x) => x.type !== 'row')?.scopedVars?.app.value).toBe('se1');
-
-      model.collapseRows();
-      const savedModelWithCollapsedRows: any = model.getSaveModelClone();
-      expect(savedModelWithCollapsedRows.panels[0].panels.length).toBe(2);
     });
   });
